refactor(console): move form-listeners component into api component folder
