+++ gravitee-apim-console-webui/src/management/api/component/gio-form-listeners/gio-form-listeners-virtual-host/gio-form-listeners-virtual-host.module.spec.ts
@@ -24,8 +24,8 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { GioFormListenersVirtualHostModule } from './gio-form-listeners-virtual-host.module';
 import { GioFormListenersVirtualHostHarness } from './gio-form-listeners-virtual-host.harness';
 
-import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../shared/testing';
-import { AjsRootScope } from '../../../ajs-upgraded-providers';
+import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../../shared/testing';
+import { AjsRootScope } from '../../../../../ajs-upgraded-providers';
 
 @Component({
   template: ` <gio-form-listeners-virtual-host [formControl]="formControl"></gio-form-listeners-virtual-host> `,
