+++ gravitee-apim-console-webui/src/management/api/component/gio-form-listeners/gio-form-listeners.common.scss
